+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Ver. 0.5.3 - Development
 ==============================
+* Added ``rule`` attribute to wsgi Router.
 * Critical bug fix in python Http parser (4bd8a54_).
 * Bug fix in :class:`pulsar.apps.wsgi.Router` metaclass.
 * **469 regression tests**, **88% coverage**.